chronos-nazi-mode-proof

+++ format/ft_pfconv_char.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_pfconv_char.c                                   :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: map <marvin@student.42.fr>                 +#+  +:+       +#+        */
+/*   By: map <map@student.42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/10/04 22:58:48 by map               #+#    #+#             */
-/*   Updated: 2016/10/06 04:47:16 by qloubier         ###   ########.fr       */
+/*   Updated: 2016/10/09 22:06:30 by qloubier         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,8 +21,8 @@ void			ft_pfconv_char(t_pfb *b, t_pfc *arg, size_t len)
 	if (arg->flag & (PFF_LONG | PFF_LONG_LONG | PFF_CAPITAL))
 	{
 		cb = (wchar_t)(arg->arg);
-		ft_printf_bwritew(b, &cb, arg->b_len);
+		ft_printf_bwritew(b, &cb, (size_t)arg->b_len);
 	}
 	else
-		ft_printf_bwritec(b, (char)(arg->arg), arg->b_len);
+		ft_printf_bwritec(b, (char)(arg->arg), (size_t)arg->b_len);
 }
