+++ format/ft_pfconv_str.c
@@ -6,7 +6,7 @@
 /*   By: qloubier <qloubier@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/10/04 19:44:05 by qloubier          #+#    #+#             */
-/*   Updated: 2016/10/06 03:06:07 by map              ###   ########.fr       */
+/*   Updated: 2016/10/09 22:12:53 by qloubier         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,10 +22,10 @@ void			ft_pfconv_str(t_pfb *b, t_pfc *arg, size_t len)
 	if (arg->arg)
 	{
 		if (arg->type == PFT_STR)
-			ft_printf_bwrite(b, (char *)(arg->arg), arg->b_len);
+			ft_printf_bwrite(b, (char *)(arg->arg), (size_t)arg->b_len);
 		else
-			ft_printf_bwritew(b, (wchar_t *)(arg->arg), arg->b_len);
+			ft_printf_bwritew(b, (wchar_t *)(arg->arg), (size_t)arg->b_len);
 	}
 	else
-		ft_printf_bwrite(b, g_pf_nullstr, arg->b_len);
+		ft_printf_bwrite(b, g_pf_nullstr, (size_t)arg->b_len);
 }
