+++ format/ft_pflen_char.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_pflen_char.c                                    :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: qloubier <marvin@student.42.fr>            +#+  +:+       +#+        */
+/*   By: qloubier <qloubier@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/10/01 02:43:11 by qloubier          #+#    #+#             */
-/*   Updated: 2016/10/06 04:39:27 by qloubier         ###   ########.fr       */
+/*   Updated: 2016/10/09 22:13:48 by qloubier         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,6 +20,6 @@ int		ft_pflen_char(t_pfc *arg)
 		arg->b_len = 1;
 	if ((arg->flag & PFF_ZERO_FILL) && !(arg->flag & PFF_LEFT_ALIGN)
 		&& (arg->minwidth))
-		return (arg->minwidth);
+		return ((int)arg->minwidth);
 	return (1);
 }
