+++ format/ft_pflen_nbr.c
@@ -6,7 +6,7 @@
 /*   By: qloubier <qloubier@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/10/01 02:43:11 by qloubier          #+#    #+#             */
-/*   Updated: 2016/10/08 20:58:24 by qloubier         ###   ########.fr       */
+/*   Updated: 2016/10/09 22:15:15 by qloubier         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 int					ft_pflen_nbr(t_pfc *arg)
 {
 	const int		neg = (arg->arg & (1ul << 63)) ? 1 : 0;
-	const uintmax_t	n = (uintmax_t)((neg) ? -(intmax_t)arg->arg : arg->arg);
+	const uintmax_t	n = (neg) ? (uintmax_t)(-(intmax_t)arg->arg) : arg->arg;
 	t_ui			len;
 
 	len = FT_MX_FLOATLEN;
@@ -23,7 +23,7 @@ int					ft_pflen_nbr(t_pfc *arg)
 		len = 0;
 	else
 		arg->tmpb = ft_ujfillbuf(n, arg->tmpb, (int *)&len);
-	arg->b_len = len;
+	arg->b_len = (int)len;
 	if ((arg->flag & PFF_PREC_SET) && (len < arg->precision))
 		len = arg->precision;
 	if ((arg->flag & (PFF_FORCE_SIGN | PFF_SPACE)) || (neg))
@@ -32,5 +32,5 @@ int					ft_pflen_nbr(t_pfc *arg)
 		&& !(arg->flag & (PFF_LEFT_ALIGN | PFF_PREC_SET))
 		&& (len < arg->minwidth))
 		len = arg->minwidth;
-	return (len);
+	return ((int)len);
 }
