+++ format/ft_printf_bflush.c
@@ -6,7 +6,7 @@
 /*   By: qloubier <qloubier@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/09/29 23:29:51 by qloubier          #+#    #+#             */
-/*   Updated: 2016/10/06 02:50:22 by map              ###   ########.fr       */
+/*   Updated: 2016/10/09 22:18:02 by qloubier         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,11 +19,11 @@ int				ft_printf_bflush(t_pfb *b)
 
 	if (!(b->blen))
 		return (0);
-	i = write(b->fd, b->buffer, b->blen);
+	i = write(b->fd, b->buffer, (size_t)b->blen);
 	if (i < 0)
-		return (i);
+		return ((int)i);
 	b->len += i;
 	b->blen = 0;
 	b->c = b->buffer;
-	return (i);
+	return ((int)i);
 }
