+++ str/ft_strchr.c
@@ -1,12 +1,12 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   strchr.c                                           :+:      :+:    :+:   */
+/*   ft_strchr.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: qloubier <marvin@42.fr>                    +#+  +:+       +#+        */
+/*   By: qloubier <qloubier@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2015/11/26 15:52:23 by qloubier          #+#    #+#             */
-/*   Updated: 2015/12/22 12:24:31 by qloubier         ###   ########.fr       */
+/*   Updated: 2016/10/09 21:45:57 by qloubier         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,8 +17,8 @@ char	*ft_strchr(const char *s, int c)
 	while (*s)
 	{
 		if (*s == (char)c)
-			return ((char *)s);
+			return ((char *)(unsigned long)s);
 		s++;
 	}
-	return (c ? NULL : (char *)s);
+	return (c ? NULL : (char *)(unsigned long)s);
 }
