+++ str/ft_strchrany.c
@@ -6,7 +6,7 @@
 /*   By: qloubier <qloubier@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/01/10 04:46:59 by qloubier          #+#    #+#             */
-/*   Updated: 2016/01/25 14:47:19 by qloubier         ###   ########.fr       */
+/*   Updated: 2016/10/09 21:46:15 by qloubier         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,10 +17,10 @@ char	*ft_strchrany(const char *s, const char *clst, char end)
 	while (*s)
 	{
 		if (ft_strchr(clst, (int)(*s)))
-			return ((char *)s);
+			return ((char *)(unsigned long)s);
 		s++;
 	}
 	if (end)
-		return ((char *)s);
+		return ((char *)(unsigned long)s);
 	return (NULL);
 }
