+++ str/ft_strrchr.c
@@ -1,12 +1,12 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   strrchr.c                                          :+:      :+:    :+:   */
+/*   ft_strrchr.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: qloubier <marvin@42.fr>                    +#+  +:+       +#+        */
+/*   By: qloubier <qloubier@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2015/11/26 16:03:52 by qloubier          #+#    #+#             */
-/*   Updated: 2015/12/18 20:41:36 by qloubier         ###   ########.fr       */
+/*   Updated: 2016/10/09 21:50:46 by qloubier         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,7 @@ char	*ft_strrchr(const char *s, int c)
 {
 	char	*as;
 
-	as = (char *)s;
+	as = (char *)(unsigned long)s;
 	while (*as)
 		(as++);
 	while (as >= s)
