+++ str/ft_strstr.c
@@ -1,12 +1,12 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   strstr.c                                           :+:      :+:    :+:   */
+/*   ft_strstr.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: qloubier <marvin@42.fr>                    +#+  +:+       +#+        */
+/*   By: qloubier <qloubier@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2015/11/26 16:27:08 by qloubier          #+#    #+#             */
-/*   Updated: 2015/12/22 17:15:47 by qloubier         ###   ########.fr       */
+/*   Updated: 2016/10/09 21:53:42 by qloubier         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,11 +18,11 @@ char	*ft_strstr(const char *s1, const char *s2)
 
 	len = ft_strlen(s2);
 	if (!len)
-		return ((char *)s1);
+		return ((char *)(unsigned long)s1);
 	while (*s1)
 	{
 		if (*s1 == *s2 && ft_strncmp(s1 + 1, s2 + 1, len - 1) == 0)
-			return ((char *)s1);
+			return ((char *)(unsigned long)s1);
 		s1++;
 	}
 	return (NULL);
