+++ unicode/ft_utf8wcsalign.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_utf8wcsalign.c                                  :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: map <marvin@student.42.fr>                 +#+  +:+       +#+        */
+/*   By: map <map@student.42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/10/06 01:31:57 by map               #+#    #+#             */
-/*   Updated: 2016/10/06 04:17:29 by qloubier         ###   ########.fr       */
+/*   Updated: 2016/10/09 22:33:43 by qloubier         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,7 +20,7 @@ size_t		ft_utf8wcsalign(const wchar_t *s, size_t mxlen)
 	len = 0;
 	while (*s && (len < mxlen))
 	{
-		i = len + ft_utf8wcsize(*(s++));
+		i = len + (size_t)ft_utf8wcsize(*(s++));
 		if (i > mxlen)
 			break ;
 		len = i;
